refactor(web): extract judge MultiSelectBar into pages/judge/MultiSelectBar.tsx (PRD #1049)
